Problems 2-4

+++ src/com/lunatech/euler/model/NaturalNumber.java
@@ -1,16 +1,16 @@
 package com.lunatech.euler.model;
 
 public class NaturalNumber {
-	private final int value;
+	private final long value;
 
-	public NaturalNumber(final int value) {
+	public NaturalNumber(final long value) {
 		if (value < 1) {
 			throw new IllegalArgumentException("Not a natural number: " + value);
 		}
 		this.value = value;
 	}
 
-	public int getValue() {
+	public long getValue() {
 		return value;
 	}
 }
